rename plots.py
