+++ pystan/plots/__init__.py
@@ -1,2 +1,2 @@
-from .plots import traceplot, forestplot, parcoords
+from .plots_plotting import traceplot, forestplot, parcoords
 from .plots_data import traceplot_data, forestplot_data, parcoords_data, kde_data, hist_data
